Так как google-service.json ушел в удаленный репо, пришлось изменить package name, чтобы изменился состав файла
