+++ app/src/androidTest/java/project/paveltoy/noteapp/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package learn.geekbrains.noteapp;
+package project.paveltoy.noteapp;
 
 import android.content.Context;
