+++ app/src/main/java/project/paveltoy/noteapp/Note.java
@@ -1,4 +1,4 @@
-package learn.geekbrains.noteapp;
+package project.paveltoy.noteapp;
 
 import android.os.Parcel;
 import android.os.Parcelable;
