+++ app/src/main/java/project/paveltoy/noteapp/RootLandFragment.java
@@ -1,7 +1,6 @@
-package learn.geekbrains.noteapp;
+package project.paveltoy.noteapp;
 
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
