+++ app/src/test/java/project/paveltoy/noteapp/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package learn.geekbrains.noteapp;
+package project.paveltoy.noteapp;
 
 import org.junit.Test;
